chore: move image attachment test to `attachments.test.ts`
